updated to nom 5.0.1
They're doing away with macros, so eventually I'll have to convert to the new way of doing things, but the macros are the reason I liked nom to begin with.  It was less work than writing functions and I understood how to chain them togethor.  Just glancing at the new way, I wouldn't know where to begin converting them to functions.