RYO SystemData is Good!
derive(SystemData) doesn't work without shred and shred_derive in cargo file whether I extern crate them or not.  Would expect it to be in specs_derive.  Anyway, this works.

The new DrawPlatformData streamlines DrawPlatform with the benefit of making it easier to read.  Looks less hacky, too :)

"Do not! ... call me a hack."
- Sealab 2021